sm8320-common: overlay: Don't pin camera in memory
The default camera app can be *huge* in some cases, e.g. when the app in
question is Google Camera. The system will only pin up to the first 80
MiB of the APK file, as well as the first 80 MiB of its odex. There are
several problems with this:

  - We could easily end up with 160 MiB of camera app files pinned,
    which is a somewhat tall order with the ~5.3 GiB of usable RAM that
    we have
  - The data that gets pinned may not even be the most critical data for
    launching the camera

Disable pinning of the camera app to save precious RAM on this device.

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>